Add files via upload
A Bank accepts deposits from customers and from the corpus thus available, it lends to Borrowers who want to carry out certain Business activities for Growth and Profit. It is often seen that due to some reasons like failure of Business, the company making losses or the company becoming delinquent/bankrupt the loans are either not Paid in full or are Charged-off or are written off. The Bank is thus faced with the problem of identifying those Borrowers who can pay up in full and not lending to borrowers who are likely to default.
At the time of giving loan to a new customer or a company, the bank needs to know the chance of getting risk on loan repayment, based on some relevant information.
Predicting the chance of risk is more relevant before giving the loan. So, it reduces the risk of profit & loss of the bank. 
When the prediction is showing high risk, the bank have to collect more information about the background of that particular company. Then bank will decide to give the loan or not.
So, our team built a model which could predict where a customer will default on their loan, using python for modelling and flask for deployment.